# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Manage a WireGuard VPN concentrator (interfaces \u0026 peers) without touching the CLI.\nKeys, client configs and QR codes are generated server-side; peers are pushed to the kernel via netlink.",
        "title": "VPN Manager API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/interfaces": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interfaces"
                ],
                "summary": "List interfaces",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interfaces"
                ],
                "summary": "Create an interface",
                "parameters": [
                    {
                        "description": "Interface payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateInterfaceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/interfaces/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interfaces"
                ],
                "summary": "Get an interface (with peers)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Interface ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interfaces"
                ],
                "summary": "Update an interface",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Interface ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Interface payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateInterfaceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interfaces"
                ],
                "summary": "Delete an interface",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Interface ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/interfaces/{id}/status": {
            "get": {
                "description": "Returns peers enriched with live handshake/transfer stats",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interfaces"
                ],
                "summary": "Interface runtime status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Interface ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/interfaces/{id}/sync": {
            "post": {
                "description": "(Re)create the link and push the current peer set to WireGuard",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interfaces"
                ],
                "summary": "Apply interface to the kernel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Interface ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {},
                "success": {
                    "type": "boolean"
                },
                "warning": {
                    "description": "Warning is set when the primary operation succeeded and was persisted, but a\nsecondary best-effort step degraded (e.g. the WireGuard kernel reconcile or\na VPN runtime apply failed). The status stays 2xx because the resource change\nwas committed; the warning carries the degradation detail for the client.",
                    "type": "string"
                }
            }
        },
        "dto.CreateInterfaceRequest": {
            "type": "object",
            "required": [
                "address",
                "endpoint",
                "listen_port",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "10.8.0.1/24"
                },
                "dns": {
                    "type": "string",
                    "maxLength": 128,
                    "example": "1.1.1.1"
                },
                "egress_interface": {
                    "description": "EgressInterface is the WAN interface for NAT; auto-detected when empty.",
                    "type": "string",
                    "maxLength": 32,
                    "example": "eth0"
                },
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "endpoint": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3,
                    "example": "vpn.example.com"
                },
                "listen_port": {
                    "type": "integer",
                    "maximum": 65535,
                    "minimum": 1,
                    "example": 51820
                },
                "masquerade": {
                    "description": "Masquerade enables NAT + forwarding so clients get internet access.",
                    "type": "boolean",
                    "example": true
                },
                "mtu": {
                    "type": "integer",
                    "maximum": 9000,
                    "minimum": 576,
                    "example": 1420
                },
                "name": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2,
                    "example": "wg0"
                },
                "private_key": {
                    "type": "string",
                    "example": ""
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.UpdateInterfaceRequest": {
            "type": "object",
            "required": [
                "address",
                "endpoint",
                "listen_port"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "10.8.0.1/24"
                },
                "dns": {
                    "type": "string",
                    "maxLength": 128,
                    "example": "1.1.1.1"
                },
                "egress_interface": {
                    "type": "string",
                    "maxLength": 32,
                    "example": "eth0"
                },
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "endpoint": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3,
                    "example": "vpn.example.com"
                },
                "listen_port": {
                    "type": "integer",
                    "maximum": 65535,
                    "minimum": 1,
                    "example": 51820
                },
                "masquerade": {
                    "type": "boolean",
                    "example": true
                },
                "mtu": {
                    "type": "integer",
                    "maximum": 9000,
                    "minimum": 576,
                    "example": 1420
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the token from /auth/login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}